Score screen back button
-> update a score screen
-> Score is almost done remain only write to json file and print it on screen

+++ Assets/Scripts/Score.cs
@@ -4,9 +4,9 @@
 public class Score
 {
     public string name;
-    public float score;
+    public int score;
 
-    public Score(string name, float score)
+    public Score(string name, int score)
     {
         this.name = name;
         this.score = score;
